Add support for custom javascript on the documentation page
In order to modify the behavior of the documentation page, it would be
nice to be able to include custom js to the existing page without
copying and changing it wholesale.

This adds a parameter, customJS, which should reply with javascript
(this is empty by default); this javascript will be included with the
documentation page.